requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==0.19.1
gunicorn==21.2.0
werkzeug==2.0.0
